+++ src/main/java/shop/zip/travel/presentation/member/MemberMyPageController.java
@@ -16,6 +16,7 @@ import shop.zip.travel.domain.member.dto.response.MemberInfoRes;
 import shop.zip.travel.domain.member.service.MemberMyPageService;
 import shop.zip.travel.domain.post.travelogue.dto.res.TravelogueCustomSlice;
 import shop.zip.travel.domain.post.travelogue.dto.res.TravelogueSimpleRes;
+import shop.zip.travel.domain.post.travelogue.service.TravelogueTempService;
 import shop.zip.travel.global.security.UserPrincipal;
 
 @RestController
@@ -25,14 +26,17 @@ public class MemberMyPageController {
   private static final int DEFAULT_SIZE = 5;
 
   private final MemberMyPageService memberService;
+  private final TravelogueTempService travelogueTempService;
 
-  public MemberMyPageController(MemberMyPageService memberService) {
+  public MemberMyPageController(MemberMyPageService memberService,
+      TravelogueTempService travelogueTempService) {
     this.memberService = memberService;
+    this.travelogueTempService = travelogueTempService;
   }
 
   @GetMapping("/info")
   public ResponseEntity<MemberInfoRes> getMyInfo(
-    @AuthenticationPrincipal UserPrincipal userPrincipal
+      @AuthenticationPrincipal UserPrincipal userPrincipal
   ) {
     MemberInfoRes memberInfo = memberService.getInfoBy(userPrincipal.getUserId());
 
@@ -70,6 +74,16 @@ public class MemberMyPageController {
         memberService.getMyBookmarkedList(userPrincipal.getUserId(), pageable);
 
     return ResponseEntity.ok(travelogues);
+  }
+
+  @GetMapping("/temp-travelogues")
+  public ResponseEntity<TravelogueCustomSlice<TravelogueSimpleRes>> getTempAll(
+      @PageableDefault(size = DEFAULT_SIZE) Pageable pageable,
+      @AuthenticationPrincipal UserPrincipal userPrincipal) {
+    TravelogueCustomSlice<TravelogueSimpleRes> travelogueSimpleResList =
+        travelogueTempService.getMyTempTravelogues(userPrincipal.getUserId(), pageable);
+
+    return ResponseEntity.ok(travelogueSimpleResList);
   }
 
 }
