+++ src/test/java/shop/zip/travel/presentation/travelogue/TravelogueControllerTest.java
@@ -45,7 +45,7 @@ import shop.zip.travel.domain.post.subTravelogue.repository.SubTravelogueReposit
 import shop.zip.travel.domain.post.travelogue.DummyGenerator;
 import shop.zip.travel.domain.post.travelogue.data.Cost;
 import shop.zip.travel.domain.post.travelogue.data.Period;
-import shop.zip.travel.domain.post.travelogue.dto.req.TravelogueCreateReq;
+import shop.zip.travel.domain.post.travelogue.dto.req.TempTravelogueCreateReq;
 import shop.zip.travel.domain.post.travelogue.entity.Travelogue;
 import shop.zip.travel.domain.post.travelogue.repository.TravelogueRepository;
 import shop.zip.travel.global.config.QuerydslConfig;
@@ -131,104 +131,6 @@ class TravelogueControllerTest {
   }
 
   @Test
-  @DisplayName("자신이 작성 중이던 임시 저장 글들을 불러올 수 있다.")
-  public void test_get_all_temp_travelogue() throws Exception {
-
-    String token = "Bearer " + jwtTokenProvider.createAccessToken(member.getId());
-
-    mockMvc.perform(get("/api/travelogues/temp")
-            .header("AccessToken", token)
-            .queryParam("size", "2")
-            .queryParam("page", "0"))
-        .andExpect(status().isOk())
-        .andDo(print())
-        .andDo(document("get-all-temp-travelogue",
-            responseFields(
-                fieldWithPath("content[]").description("").optional(),
-                fieldWithPath("content[].travelogueId").type(JsonFieldType.NUMBER)
-                    .description("Travelogue pk"),
-                fieldWithPath("content[].title").type(JsonFieldType.STRING)
-                    .description("Travelogue 제목"),
-                fieldWithPath("content[].nights").type(JsonFieldType.NUMBER).description("숙박 일"),
-                fieldWithPath("content[].days").type(JsonFieldType.NUMBER).description("여행 전체 일"),
-                fieldWithPath("content[].totalCost").type(JsonFieldType.NUMBER)
-                    .description("여행 전체 비용"),
-                fieldWithPath("content[].country").type(JsonFieldType.STRING).description("방문한 나라"),
-                fieldWithPath("content[].thumbnail").type(JsonFieldType.STRING)
-                    .description("썸네일 링크"),
-                fieldWithPath("content[].member.nickname").type(JsonFieldType.STRING)
-                    .description("작성자 닉네임"),
-                fieldWithPath("content[].member.profileImageUrl").type(JsonFieldType.STRING)
-                    .description("작성자 프로필 이미지 링크"),
-                fieldWithPath("pageable.sort.empty").type(JsonFieldType.BOOLEAN)
-                    .description("데이터가 비어있는지에 대한 여부"),
-                fieldWithPath("pageable.sort.sorted").type(JsonFieldType.BOOLEAN)
-                    .description("데이터가 정렬되어있는지에 대한 여부"),
-                fieldWithPath("pageable.sort.unsorted").type(JsonFieldType.BOOLEAN)
-                    .description("데이터가 정렬되어 있지 않은지에 대한 여부"),
-                fieldWithPath("pageable.offset").type(JsonFieldType.NUMBER)
-                    .description("페이징 offset"),
-                fieldWithPath("pageable.pageNumber").type(JsonFieldType.NUMBER)
-                    .description("현재 요청한 페이지 넘버"),
-                fieldWithPath("pageable.pageSize").type(JsonFieldType.NUMBER)
-                    .description("요청한 데이터 갯수"),
-                fieldWithPath("pageable.paged").type(JsonFieldType.BOOLEAN)
-                    .description("페이징이 된 여부"),
-                fieldWithPath("pageable.unpaged").type(JsonFieldType.BOOLEAN)
-                    .description("페이징이 되지 않은 여부"),
-                fieldWithPath("size").type(JsonFieldType.NUMBER).description("요청된 페이징 사이즈"),
-                fieldWithPath("number").type(JsonFieldType.NUMBER).description("페이지 번호"),
-                fieldWithPath("numberOfElements").type(JsonFieldType.NUMBER)
-                    .description("조회된 데이터 갯수"),
-                fieldWithPath("first").type(JsonFieldType.BOOLEAN).description("첫번째 페이지인지의 여부"),
-                fieldWithPath("last").type(JsonFieldType.BOOLEAN).description("마지막 페이지인지의 여부"),
-                fieldWithPath("empty").type(JsonFieldType.BOOLEAN).description("데이터가 없는지의 여부")
-            )));
-  }
-
-  @Test
-  @DisplayName("메인 게시물을 저장할 수 있다.")
-  void test_save_travelogue() throws Exception {
-    // given
-    TravelogueCreateReq travelogueCreateReq = new TravelogueCreateReq(
-        DummyGenerator.createPeriod(),
-        "메인 게시물 제목",
-        DummyGenerator.createCountry(),
-        "www.naver.com",
-        DummyGenerator.createCost()
-    );
-
-    String token = "Bearer " + jwtTokenProvider.createAccessToken(member.getId());
-
-    mockMvc.perform(post("/api/travelogues")
-            .header("AccessToken", token)
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(objectMapper.registerModule(new JavaTimeModule())
-                .writeValueAsString(travelogueCreateReq)))
-        .andExpect(status().isOk())
-        .andDo(print())
-        .andDo(document("save-travelogue",
-            requestFields(
-                fieldWithPath("period.startDate").description("여행 시작 날짜"),
-                fieldWithPath("period.endDate").description("여행 종료 날짜"),
-                fieldWithPath("period.nights").description("여행 숙박 횟수"),
-                fieldWithPath("title").description("게시물 제목"),
-                fieldWithPath("country.name").description("여행한 나라 이름"),
-                fieldWithPath("thumbnail").description("게시물 썸네일 URL"),
-                fieldWithPath("cost.transportation").description("이동 수단 경비"),
-                fieldWithPath("cost.lodge").description("숙박 비용"),
-                fieldWithPath("cost.etc").description("기타 비용"),
-                fieldWithPath("cost.total").description("전체 경비")
-            ),
-            responseFields(
-                fieldWithPath("id").description("생성된 게시물의 pk 값"),
-                fieldWithPath("nights").description("n박"),
-                fieldWithPath("days").description("n일")
-
-            )));
-  }
-
-  @Test
   @DisplayName("임시 저장된 게시글을 발행할 수 있다.")
   void test_publish_travelogue() throws Exception {
 
@@ -264,20 +166,20 @@ class TravelogueControllerTest {
   }
 
   @Test
-  @DisplayName("값들이 비어있는 게시물을 임시 저장할 수 있다.")
+  @DisplayName("메인 게시글을 작성 혹은 임시 작성 할 수 있다.")
   void test_temp_save_travelogue() throws Exception {
     // given
-    TravelogueCreateReq travelogueCreateReq = new TravelogueCreateReq(
-        DummyGenerator.createPeriod(),
+    TempTravelogueCreateReq travelogueCreateReq = new TempTravelogueCreateReq(
+        DummyGenerator.createTempPeriod(),
         null,
-        DummyGenerator.createCountry(),
+        DummyGenerator.createTempCountry(),
         "www.naver.com",
-        DummyGenerator.createCost()
+        DummyGenerator.createTempCost()
     );
 
     String token = "Bearer " + jwtTokenProvider.createAccessToken(member.getId());
 
-    mockMvc.perform(post("/api/travelogues/temp")
+    mockMvc.perform(post("/api/travelogues")
             .header("AccessToken", token)
             .contentType(MediaType.APPLICATION_JSON)
             .content(objectMapper.registerModule(new JavaTimeModule())
